+++ src/Form/FormControlTypes.php
@@ -7,6 +7,7 @@ namespace CraftCms\Cms\Form;
 use CraftCms\Cms\Component\TypeRegistry;
 use CraftCms\Cms\Form\Contracts\Control;
 use CraftCms\Cms\Form\Controls\Address;
+use CraftCms\Cms\Form\Controls\AssetSelect;
 use CraftCms\Cms\Form\Controls\Checkbox;
 use CraftCms\Cms\Form\Controls\Choice;
 use CraftCms\Cms\Form\Controls\Color;
@@ -50,6 +51,7 @@ class FormControlTypes extends TypeRegistry
 
     protected const array DEFAULT_TYPES = [
         Address::class,
+        AssetSelect::class,
         Checkbox::class,
         Choice::class,
         ConditionBuilder::class,